config/api_router.py
from django.conf import settings
from rest_framework.routers import DefaultRouter, SimpleRouter
from django.urls import path
from rest_framework_simplejwt.views import TokenRefreshView, TokenObtainPairView


# Local Imports
from apps.users.api.views import UserViewSet

if settings.DEBUG:
    router = DefaultRouter()
else:
    router = SimpleRouter()

# USERS
router.register("users", UserViewSet, basename="user")




app_name = "api"
urlpatterns = [
    path("token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
]
urlpatterns += router.urls